Moved some common SQL function tu public utls as they are used in plugins
