+++ server/channels/store/sqlstore/utils.go
@@ -183,31 +183,6 @@ func AppendBinaryFlag(buf []byte) []byte {
 	return append([]byte{0x01}, buf...)
 }
 
-// AppendMultipleStatementsFlag attached dsn parameters to MySQL dsn in order to make migrations work.
-func AppendMultipleStatementsFlag(dataSource string) (string, error) {
-	config, err := mysql.ParseDSN(dataSource)
-	if err != nil {
-		return "", err
-	}
-
-	if config.Params == nil {
-		config.Params = map[string]string{}
-	}
-
-	config.Params["multiStatements"] = "true"
-	return config.FormatDSN(), nil
-}
-
-// ResetReadTimeout removes the timeout constraint from the MySQL dsn.
-func ResetReadTimeout(dataSource string) (string, error) {
-	config, err := mysql.ParseDSN(dataSource)
-	if err != nil {
-		return "", err
-	}
-	config.ReadTimeout = 0
-	return config.FormatDSN(), nil
-}
-
 func SanitizeDataSource(driverName, dataSource string) (string, error) {
 	switch driverName {
 	case model.DatabaseDriverPostgres:
